requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Werkzeug==2.3.7
python-dotenv==1.0.1
pillow==10.2.0
pyzbar==0.1.9
openpyxl==3.1.2
pandas==2.2.1
gunicorn
psycopg
SQLAlchemy==2.0.25
